feat: InputLock component
