+++ packages/extension-ui/src/components/index.ts
@@ -52,6 +52,7 @@ export { default as HelperFooter } from './HelperFooter';
 export { default as ScrollWrapper } from './ScrollWrapper';
 export { default as BottomWrapper } from './BottomWrapper';
 export { default as WarningBox } from './WarningBox';
+export { default as InputLock } from './InputLock';
 
 export { Input, TextArea } from './TextInputs';
